Overhaul packaging, linting, and dependencies
This migrates this project to a fully pyproject.toml based project
following current best practices. Linting, formatting, and typechecking
have been replaced with Ruff and mypy, as Ruff is currently the only
really well maintained linter/formatter. Upgrading mypy versions it
found a couple more errors so I quickly corrected them here.

The dependencies have been pruned and updated. The big one is pulling in
the LOST python interface directly from github since it is not on pypi.
I've also updated olaf to the latest and removed the fixed version on
opencv. There was no description given as to why the opencv version was
fixed and a quick smoke test works so I assume we're fine to upgrade.
This will also stop Dependabot spamming about it.

This also fixes CI to install the right (pypi) version of
oresat-configs instead of directly from github, as it always should have
been.